Hydra-package {Hydra}	R Documentation

Hydra: Hydrating Package Skeletons into Executable R Study Packages

Description

An R package and Java library for hydrating package skeletons into executable R study packages based on specifications in JSON format.

Author(s)

Maintainer: Anthony Sena sena@ohdsi.org

Authors:

Martijn Schuemie schuemie@ohdsi.org

Jenna Reps

See Also

Useful links:

https://ohdsi.github.io/Hydra

https://github.com/OHDSI/Hydra

Report bugs at https://github.com/OHDSI/Hydra/issues

[Package Hydra version 0.3.0 Index]