hydrate {Hydra}	R Documentation

Hydrate a package skeleton according to study specifications

Description

Hydrate a package skeleton according to study specifications

Usage

hydrate(
  specifications,
  outputFolder,
  skeletonFileName = NULL,
  packageName = NULL
)

Arguments

specifications
A string containing JSON specifications for hydrating the study. Specifications will be generated by ATLAS.

outputFolder
The folder where the hydrated package will be stored.

skeletonFileName
(Mostly for debugging purposes:) Specify the location of a package skeleton to hydrate. If not specified, the appropriate skeleton will be selected from those inside the Hydra package.

packageName
The name of the R package to create. If provided, will override the packageName variable in the specifications JSON.

[Package Hydra version 0.3.0 Index]