prepareForOfflineStudyPackageExecution {Hydra}	R Documentation

Prepare system to run hydrated packages without further internet connection

Description

Prepare system to run hydrated packages without further internet connection

Usage

prepareForOfflineStudyPackageExecution(
  installRpackages = TRUE,
  installJdbcDrivers = TRUE,
  skeletons = listSkeletons()
)

Arguments

installRpackages
Install the R packages required by the skeletons?

installJdbcDrivers
Install all JDBC drivers? Requires the DATABASECONNECTOR_JAR_FOLDER environmental variable to be set.

skeletons
A list of skeletons to check, for example 'CohortDiagnosticsStudy_v0.0.1.zip'.

Details

Note that when installJdbcDrivers = TRUE this will only include the Jar drivers supported by DatabaseConnector::downloadJdbcDrivers. Other drivers (like BigQuery) will need to be downloaded manually and placed in the folder identified by the DATABASECONNECTOR_JAR_FOLDER environmental variable.

Use list.files(system.file("skeletons", package = "Hydra"), pattern = "*.zip") for a list of all skeletons

Value

This function does not return anything. Instead, it installs all dependencies required to run the hydrated skeletons. (Only those skeletons that use a renv.lock file)

[Package Hydra version 0.3.0 Index]